requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
bleach 
flask-swagger-ui
python-dotenv